requirements.txt
mlflow==2.0.0
psycopg2-binary==2.9.0
requests==2.28.0
scikit-learn==1.0.0
pandas==1.5.0
numpy==1.23.0
pytest==7.2.0
slack_sdk==3.20.0
python-dotenv==0.21.0
evidently==0.4.0
xgboost>=1.7.0
shap>=0.41.0